fix(build): reject unsupported CommonJS requests
